modif shaders

+++ wolf3d/structures.h
@@ -6,7 +6,7 @@
 /*   By: edescoin <edescoin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/03 16:19:46 by edescoin          #+#    #+#             */
-/*   Updated: 2017/05/13 20:58:26 by edescoin         ###   ########.fr       */
+/*   Updated: 2017/05/15 23:17:53 by edescoin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -147,7 +147,7 @@ typedef struct		s_mob
 typedef struct		s_map_data
 {
 	int				brightness;
-	int				reflection;
+	double			reflection;
 	double			bg_fact;
 	t_texture		*bgd;
 	SDL_Color		floor;
